+++ targets/core/nxp/lpc175x/power.hpp
@@ -1,9 +1,9 @@
-#ifndef KLIB_LPC1756_POWER_CONTROL_HPP
-#define KLIB_LPC1756_POWER_CONTROL_HPP
+#ifndef KLIB_NXP_LPC175X_POWER_CONTROL_HPP
+#define KLIB_NXP_LPC175X_POWER_CONTROL_HPP
 
 #include <lpc1756.hpp>
 
-namespace klib::lpc1756 {
+namespace klib::core::lpc175x::io {
     class power_control {
     public:
         /**
