+++ targets/core/nxp/lpc175x/rtc.hpp
@@ -1,31 +1,14 @@
-#ifndef KLIB_LPC1756_RTC_HPP
-#define KLIB_LPC1756_RTC_HPP
+#ifndef KLIB_NXP_LPC175X_RTC_HPP
+#define KLIB_NXP_LPC175X_RTC_HPP
 
 #include <cstdint>
 
-#include <lpc1756.hpp>
-
+#include <klib/klib.hpp>
 #include <klib/units.hpp>
 
 #include "power.hpp"
 
-namespace klib::lpc1756::io::periph {
-    struct rtc0 {
-        // peripheral id (e.g rtc0, rtc1)
-        constexpr static uint32_t id = 0;
-
-        // peripheral clock bit position
-        constexpr static uint32_t clock_id = 9;
-
-        // interrupt id (including the arm vector table)
-        constexpr static uint32_t interrupt_id = 33;
-
-        // port to the RTC hardware
-        static inline RTC_Type *const port = RTC;
-    };
-}
-
-namespace klib::lpc1756::io {
+namespace klib::core::lpc175x::io {
     template <typename Rtc>
     class rtc {
     protected:
