+++ targets/core/nxp/lpc175x/spi.hpp
@@ -1,62 +1,17 @@
-#ifndef KLIB_LPC1756_SPI_HPP
-#define KLIB_LPC1756_SPI_HPP
+#ifndef KLIB_NXP_LPC175X_SPI_HPP
+#define KLIB_NXP_LPC175X_SPI_HPP
 
 #include <cstdint>
 
-#include <lpc1756.hpp>
-
+#include <klib/klib.hpp>
 #include <klib/io/core_clock.hpp>
 #include <klib/io/bus/spi.hpp>
 
 #include "clocks.hpp"
 #include "power.hpp"
-#include "pins.hpp"
+#include "port.hpp"
 
-namespace klib::lpc1756::io::periph::detail::spi {
-    enum class mode {
-        miso,
-        mosi,
-        sck,
-        cs0
-    };
-
-    template <typename Pin, mode Type, typename Periph>
-    struct spi {
-        // pin of the peripheral
-        using pin = Pin;
-
-        // type of the pin
-        constexpr static mode type = Type;
-
-        // alternate function
-        using periph = Periph;
-    };
-}
-
-namespace klib::lpc1756::io::periph::lqfp_80 {
-    struct spi0 {
-        // peripheral id (e.g spi0, spi1)
-        constexpr static uint32_t id = 0;
-
-        // interrupt id (including the arm vector table)
-        constexpr static uint32_t interrupt_id = 29;
-
-        // peripheral clock bit position
-        constexpr static uint32_t clock_id = 8;
-
-        // port to the SPI hardware
-        static inline SPI_Type *const port = SPI;
-
-        // configuration of the pins 
-        using mosi = detail::spi::spi<pins::package::lqfp_80::p45, detail::spi::mode::mosi, io::detail::alternate::func_3>;
-        using miso = detail::spi::spi<pins::package::lqfp_80::p46, detail::spi::mode::miso, io::detail::alternate::func_3>;
-        using sck = detail::spi::spi<pins::package::lqfp_80::p47, detail::spi::mode::sck, io::detail::alternate::func_3>;
-        using cs0 = detail::spi::spi<pins::package::lqfp_80::p48, detail::spi::mode::cs0, io::detail::alternate::func_3>;
-    };
-}
-
-
-namespace klib::lpc1756::io {
+namespace klib::core::lpc175x::io {
     template <typename Spi>
     class spi {
     protected:
