+++ targets/core/nxp/lpc175x/system.hpp
@@ -1,11 +1,10 @@
-#ifndef KLIB_LPC1756_SYSTEM_HPP
-#define KLIB_LPC1756_SYSTEM_HPP
+#ifndef KLIB_NXP_LPC175X_SYSTEM_HPP
+#define KLIB_NXP_LPC175X_SYSTEM_HPP
 
-#include <lpc1756.hpp>
-
+#include <klib/klib.hpp>
 #include <klib/io/core_clock.hpp>
 
-namespace klib::lpc1756::io::system {
+namespace klib::core::lpc175x::io::system {
     class flash {
     public:
         /**
