+++ targets/core/nxp/lpc175x/watchdog.hpp
@@ -1,24 +1,11 @@
-#ifndef KLIB_LPC1756_WATCHDOG_HPP
-#define KLIB_LPC1756_WATCHDOG_HPP
+#ifndef KLIB_NXP_LPC175X_WATCHDOG_HPP
+#define KLIB_NXP_LPC175X_WATCHDOG_HPP
 
 #include <cstdint>
 
-#include <lpc1756.hpp>
+#include <klib/klib.hpp>
 
-namespace klib::lpc1756::io::periph {
-    struct wdt0 {
-        // peripheral id (e.g wdt0, wdt1)
-        constexpr static uint32_t id = 0;
-
-        // peripheral interrupt position
-        constexpr static uint32_t interrupt_id = 16;
-
-        // port to the wdt hardware
-        static inline WDT_Type *const port = WDT;
-    };
-}
-
-namespace klib::lpc1756::io {
+namespace klib::core::lpc175x::io {
     template <typename Wdt>
     class watchdog {
     public:
